backend/requirements.txt
# ========================================
# JARVIS - Assistente de Voz Pessoal
# Requisitos Python 3.10+
# ========================================

# === CORE ===
python-dotenv>=1.0.0
pydantic>=2.5.0
colorama>=0.4.6
rich>=13.7.0

# === ÁUDIO ===
pyaudio>=0.2.14
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
scipy>=1.11.0

# === SPEECH TO TEXT (Whisper local) ===
# Usando versão mais leve que não precisa de compilação
openai-whisper>=20231117

# === WAKE WORD (Vosk) ===
vosk>=0.3.45

# === TEXT TO SPEECH ===
edge-tts>=6.1.9
pygame>=2.5.2

# === VOICE AUTHENTICATION ===
resemblyzer>=0.1.3
# webrtcvad removido (precisa compilar)

# === LLM LOCAL (Ollama) ===
ollama>=0.3.0

# === CONTROLE DO PC ===
pyautogui>=0.9.54
psutil>=5.9.7
pywin32>=306
pycaw>=20230407
comtypes>=1.2.0
keyboard>=0.13.5
pyperclip>=1.8.2
screen-brightness-control>=0.22.0

# === SERVIDOR/API (para interface) ===
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0

# === CLIMA ===
requests>=2.31.0
